Cash Register What Is Left
Given a price, cash (the amount a customer gives you), and cash in the drawer, determine if there is enough to make exact change on the transaction. If you cannot make exact change (not enough cash or not enough type of currency to make exact change) say it is insufficient. If you make exact change and have nothing left in the drawer, close the drawer. If you have enough to make exact change, keep it open and report the new cash in drawer amount.