Cash Register Change Given
When given price (cost of item), cash (cash customer gives you), and cid (amount of cash in drawer of register broken down by bills/coins), determine the status of the register after the transaction and the change given to the customer. If the change needed is exactly what the register has, change status to closed. If the change needed is greater than what's in the register or you cannot provide exact change with the currency in the register, change status to insufficient funds. If the change needed is in the register and there's still cash in the drawer after giving change, status is open and record the change given